requirements.txt
torch>=2.0.0
nvidia-ml-py>=12.0.0
PyYAML>=6.0
